Simplify getting the reviewer lists
Refactor: GitHub provides the list of requested reviewers in
the pull request event payload. This commit switches from
reading in the CODEOWNERS file, parsing it, compairing the list
of files changed in the PR to CODEOWNERS, and then requesting
reviews from any matches to simply using the list that is given
in the GH payload.

:)

Therefore, the entire utils directory is no longer needed and has
been removed.

Fixes #12